+++ CHANGELOG.md
@@ -17,5 +17,7 @@ scheme are considered to be bugs.
   the resulting state in a defined hook.
 - Buildable module to mix into modules that you want to be able to build.
 - Default state via the `Builder#defaults` extension method.
+- Including a `Buildable` module includes the default built version in the
+  descendant class or module.
 
 [unreleased]: https://github.com/michaelherold/module_builder